Fix spinner bug in album & artist view & add option to start minimized
